ARM: zero-sugar_defconfig: build CFG80211 and MAC80211 as module
On 5.4 kernel, regulatory database is loaded from firmware file
regulatory.db.  Right now, the database loading fails like below.
That's because rootfs accommodating the firmware is not ready yet when
cfg80211 starts running.

[    7.218726] cfg80211: Loading compiled-in X.509 certificates for regulatory database
[    7.229948] cfg80211: Loaded X.509 cert 'sforshee: 00b28ddf47aef9cea7'
[    7.236647] platform regulatory.0: Direct firmware load for regulatory.db failed with error -2
[    7.245333] cfg80211: failed to load regulatory.db
[    7.262905] EXT4-fs (mmcblk2p2): mounted filesystem with ordered data mode. Opts: (null)
[    7.271183] ext4 filesystem being mounted at /root supports timestamps until 2038 (0x7fffffff)
[    7.279881] VFS: Mounted root (ext4 filesystem) on device 179:2.

Build cfg80211 (and mac80211) as module, so that when it gets loaded
from rootfs, regulatory.db is ready to be requested as a firmware.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>