wacom_i2c: Revert "wacom_i2c:not disabling i2c interrupt when suspending"
This reverts commit 32017a21d99f673d62ef04d6fdb8ec56e2468ea3

This commit can be safely reverted after:

6e7860cbb80ac40f05a0ed3fafb5214309f823c9
[BACKPORT] genirq/PM: Introduce IRQCHIP_ENABLE_WAKEUP_ON_SUSPEND flag

An interrupt that is disabled/masked but set for wakeup may still need to
be able to wake up the system from sleep states like "suspend to RAM".

To that effect, introduce the IRQCHIP_ENABLE_WAKEUP_ON_SUSPEND flag.
If the irqchip have this flag set, the irq PM code will enable/unmask
the irqs that are marked for wakeup, but that are in a disabled state.

On resume, such irqs will be restored back to their disabled state.

Suggested-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Maulik Shah <mkshah@codeaurora.org>
[maz: commit message fix-up]
Signed-off-by: Marc Zyngier <maz@kernel.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Thomas Gleixner <tglx@linutronix.de>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Link: https://lore.kernel.org/r/1601267524-20199-4-git-send-email-mkshah@codeaurora.org

501a32ce44356a78f18ad0e7bc4da2a9c69b92c1
genirq/PM: Re-enable pending wake-up interrupt on suspend

A disabled level interrupt marked for wakeup and occuring before being
armed, is simply masked and will not not cause suspend aborting. This
change checks if such interrupt is pending in the irq suspend path and
re-enable it in pending case, offering a chance for the interrupt to
have it's handler executed and pm_suspend aborted.

Signed-off-by: Loic Poulain <loic.poulain@linaro.org>