input: touchscreen: synaptics_dsx: Power off on deep-suspend
Touch is not used in suspend and so can be turned off to save power.
This is already what seems to happen at electrical level, but the
reason is not clear. make power-off explicitely in the driver to
have better control of that state. Once power is reestablished,
controller asserts attention pin, and reinitialization/reconfiguration
is done from the interrupt handler.

Signed-off-by: Loic Poulain <loic.poulain@linaro.org>