max77818: add handling of SOC alerts
SOC alerts are cleared by setting new alert threshold values.
SOC alerts are disabled by default and can be configured from the device tree.
If configured in the device tree:
During Initialization the min SOC alert is enabled and the max SOC alert
is disabled. Once the SOC drops down to or below the min alert threshold
the min SOC alert is disabled and the max soc alert in enabled.
The opposite happens when the SOC rises to or above the max alert threshold.

Signed-off-by: Michal Koziel <michal.koziel@emlogic.no>