input: touchscreen: parade: Include device access in main module
Since we now build the parade driver as modules, the pt_device_access
is built as a separated module. But this module does not depend on
any device and so is not automatically loaded at boot time.

Since it exposes features for factory testing (mfg_test), we need
to have it loaded. This patch makes the pt_device_access code part
of the main pt module so that it is loaded along with the main
code (automatically).

Alternatively, the testing application could load the pt_device_access
manually, but it's more complexity, especially if there are multiple
touch variants to deal with.

Signed-off-by: Loic Poulain <loic.poulain@linaro.org>