dts/zero-sugar: max77818: disable TAlrt interrupt
TAlrt interrupt causes the interrupt line to go low when the averege
temperatue drops below the T2 or rises above the T3 thresholds. Since
the interrupt caused by TAlrt is currently not handled, it can be
disabled. The JEITA functionality related to the T2 and T3 thresholds
is not affected.