ARM: imx: skip enet init when FEC Ethernet controller is not enabled
There is no need to initialize the enet when the FEC Ethernet controller
(of ColdFire and some i.MX CPUs) is not enabled.

This initialization is also problematic and causes hang when booting
into a new kernel with kexec. Since the enet is not used anyway with
reMarkable devices, making it CONFIG_FEC dependent allows kexec to work.

Signed-off-by: Grzegorz Jaszczyk <grzegorz.jaszczyk@linaro.org>