dts: zero-sugar: dedicated regulator for display
The VCOM output is enabled along with the SY7636A PMIC, and since we
frequently wakeup the PMIC for thermal info polling, it causes EPD
VCOM to also be enabled but without any data exchanged. This can be
prevented by decoupling VCOM output from PMIC power.

When the VCOMCTL bit is set, the VCOM output is controlled via VCOM_EN
pin, which can then be part of a dedicated regulator (controlled via
gpio).

Signed-off-by: Loic Poulain <loic.poulain@linaro.org>