dts/zero-sugar: max77818: disable max SOC alert
The consequence of this patch is that the SOC hysteresis will be
disabled and SOC min will remain set, even if SOC is below min.

Warning!
Once the min SOC threshold has been reached, max77818 will pull
the interrupt line low and keep it low as long as SOC is below
the min threshold.

Signed-off-by: Michal Koziel <michal.koziel@emlogic.no>