thermal: sy7636a: Prevent double power toggling
Thermistor requires SY7636A rails to be powered to get a valid
value (VREF), causing short toggling when thermal susbsystem
updates or polls the temperature value. It especially observed
at system resume time, a first toggling which is caused by
thermal update and a second one caused by display enabling.
Since the rails powering is a heavy and long operation it may
be worth to delay rail power-off a bit in case it is reused
shortly after. This can be done with deferred regulator
disabling.

Signed-off-by: Loic Poulain <loic.poulain@linaro.org>