max77818_battery: Always apply config2 from devicetree
Some value that are learned during lifetime should not be updated,
at each boot. However config2 register is basic configuration which
should always be aligned with devicetree.

This fix issue with non-applied devicetree config2.

We likely have the same problem with some other registers, and it
should be carefully reviewed.

Signed-off-by: Loic Poulain <loic.poulain@linaro.org>