power: suspend: Fix unbalanced sleep lock
The previously introduced workaround:
"PM: Don't sleep for 10 seconds after wakeup"

Introduce unbalanced lock holding when suspend prepare is failing,
which can happen when a kthread or a component of the suspend chain
temporarly refuses to freeze. In that case the 'resleep_mutex' is
held forever and the device will not be able to sleep until next
reboot.

Note: We may want to revisit this 'now useless'? resleep workaround
at some point.

Signed-off-by: Loic Poulain <loic.poulain@linaro.org>