wakeup_reason: fix wakeup_reason to not show impossible reasons
This fixes wakeup_reason showing the most recent wakeup source even
though it was not actually the wakeup source responsible for the wakeup.
If a wakeup happens and no registered wakeup source was responsible,
this would lead to showing the most recent registered wakeup source,
even if it wasn't actually responsible.

This change makes wakeup_reason only consider sources which were
activated after the last time the system entered suspend.

Signed-off-by: Haakon Drews <haakon.drews@remarkable.no>