sy7636a: handle failing to disable regulator
In sy7636a_get_temp we defer disabling the regulator. If a disable call
fails, it is not counted in terms of balancing enable vs. disable calls.
To make sure that enable and disable calls are balanced, only call
regulator_enable in sy7636a_get_temp if our previous regulator_disable
call was successful.

Signed-off-by: Haakon Drews <haakon.drews@remarkable.no>