pogo: consume entire incoming uart RX buffer
This adds support for processing multiple POGO messages coming in via
one receive_buf call, ie the entire RX buffer. This can happen when
we're busy and multiple RX messages are buffered before receive_buf is
called or if receive_buf aborts without processing any messages because
the FSM thread is still
busy.

To support this, fsm_routine_uart_keyboard will now iterate over the
contents of the read_fifo until now more messages can be found/parsed.

Error handling is as before. Failure to parse a header in receive_buf
will result in dismissing the rest of the RX buffer. Failure to parse a
header/message in fsm_routine_uart_keyboard will result in aborting
parsing and ultimately a timeout which will trigger a reset of read_fifo
in entry_idle.

Signed-off-by: Haakon Drews <haakon.drews@remarkable.no>