pogo: drop received data if serdev is not ready
mutex_trylock() in pogo_onewire_receive_buf() was preventing a
situation where pogo lock was taken by fsm and serdev_device_close()
was called while pogo_onewire_receive_buf() was holding the
tty lock. This would result in a deadlock if we didn't return
from pogo_onewire_receive_buf() when pogo lock was already taken.

A side effect was that the received data would not be
processed until next packet was received. This would
result in delayed processing of the first packet. This would
sometimes cause multiple keys being displayed when
a releas event was delayed.

To avoid this we postpone closing of serdev till after we
have turned VBUS off. MCU should be off by then and should
not send any data after VBUS is turned on since the driver
initiates the new session.
Additionaly, pogo_onewire_receive_buf() drops all data and returns
if a session has ended and serdev should be considered as
not ready for communication.