otgcontrol: register device on module init (#344)
The otgcontrol device has been disabled in DT since we have
started using the pogo driver. However, we still need otgcontrol
in production. We don't want the device to be enabled in DT not
to cause any conflicts with the pogo device, even if the otgcontrol
driver can be blacklisted, and only loaded on request.

This patch allows to bring up the device when the driver is loaded
and read its DT nodes as if it was enabled in DT in the first place.

Signed-off-by: Michal Koziel <michal.koziel@emlogic.no>